Modificate code

+++ app.js
@@ -1,5 +1,5 @@
-const Discord = require("discord.js");
-const bot = new Discord.Client({disableEveryone: true});
+const { Client, MessageEmbed } = require("discord.js");
+const bot = new Client({disableMentions: "all"});
 const Words = require("./words.json");
 require("dotenv").config();
 bot.login(process.env.token);
@@ -10,45 +10,45 @@ bot.on("ready", () => {
 });
 
 bot.on("guildMemberAdd", (member) => {
-    const group = bot.guilds.cache.get(member.guild.id);
+    const guild = bot.guilds.cache.get(member.guild.id);
 
-    const NewMember = new Discord.MessageEmbed()
+    const NewMember = new MessageEmbed()
         .setColor(process.env.EmbedGreen)
         .setTimestamp()
         .setThumbnail(member.user.displayAvatarURL({size: 4096, dynamic: true}))
         .setTitle(`${bot.user.username} » ${member.user.username} (${member.user.id})`)
-        .setFooter(`Участник присоединился`, group.iconURL({size: 4096, dynamic: true}));
+        .setFooter(`Участник присоединился`, guild.iconURL({size: 4096, dynamic: true}));
 
     return bot.channels.cache.get(process.env.AdminChannel).send(NewMember);
 });
 
 bot.on("guildMemberRemove", (member) => {
-    const group = bot.guilds.cache.get(member.guild.id);
+    const guild = bot.guilds.cache.get(member.guild.id);
 
-    const OldMember = new Discord.MessageEmbed()
+    const OldMember = new MessageEmbed()
         .setColor(process.env.EmbedRed)
         .setTimestamp()
         .setThumbnail(member.user.displayAvatarURL({size: 4096, dynamic: true}))
         .setTitle(`${bot.user.username} » ${member.user.username} (${member.user.id})`)
-        .setFooter(`Участник вышел`, group.iconURL({size: 4096, dynamic: true}));
+        .setFooter(`Участник вышел`, guild.iconURL({size: 4096, dynamic: true}));
 
     return bot.channels.cache.get(process.env.AdminChannel).send(OldMember);
 });
 
 bot.on("messageUpdate", (message, newMessage) => {
-    if(message.author.id == process.env.OwnerID) return;
+    if(message.author.id == process.env.OwnerID || message.member.roles.cache.has(process.env.PremiumRole)) return;
     if(Words.wh_word.some(word => newMessage.content.includes(word))) return;
     if(Words.bad_word.some(word => newMessage.content.includes(word))){
         message.delete();
 
-        const group = bot.guilds.cache.get(message.guild.id);
-        const BadWord = new Discord.MessageEmbed()
+        const guild = bot.guilds.cache.get(message.guild.id);
+        const BadWord = new MessageEmbed()
             .setColor(process.env.EmbedRed)
             .setTimestamp()
             .setThumbnail(message.author.displayAvatarURL({size: 4096, dynamic: true}))
             .setTitle(`${bot.user.username} » Запрещённое слово!`)
             .setDescription(`${message.author}, использовать такие слова запрещено!`)
-            .setFooter(`${bot.user.username}`, group.iconURL({size: 4096, dynamic: true}));
+            .setFooter(bot.user.username, guild.iconURL({size: 4096, dynamic: true}));
 
         return message.channel.send(BadWord).then(msg => msg.delete({timeout: 60000}));
     }
@@ -61,16 +61,28 @@ bot.on("message", async message => {
         message.react(process.env.ReactionDOWN);
     }
 
-    const group = bot.guilds.cache.get(message.guild.id);
-    const messages = message.content.toLowerCase().replace(/\s+/g, "");
+    const guild = bot.guilds.cache.get(message.guild.id);
+    const messages = message.content.toLowerCase();
 
     if(message.author.id === process.env.OwnerID){
+        if(message.content.toLowerCase() === "stats"){
+            message.delete();
+
+            const StatusEmbed = new MessageEmbed()
+                .setColor(process.env.EmbedGreen)
+                .setTitle(`${bot.user.username} » Статистика`)
+                .setDescription(`**set-welcome** - Сообщение приветствия.\n**set-ticket** - Сообщение для создания тикетов.\n**set-rules** - Сообщение с правилами.\n\nПинг бота: ${bot.ws.ping}`)
+                .setFooter(`JeelangaHelper`, guild.iconURL({size: 4096, dynamic: true}));
+
+            return message.channel.send(StatusEmbed).then(msg => msg.delete({timeout: 60000}));
+        }
+
         if(message.content.toLowerCase() === "set-welcome"){
             message.delete();
 
-            const WelcomeEmbed = new Discord.MessageEmbed()
+            const WelcomeEmbed = new MessageEmbed()
                 .setColor(process.env.EmbedGreen)
-                .setTitle(`**Добро пожаловать на сервер ${group.name}!**`)
+                .setTitle(`**Добро пожаловать на сервер ${guild.name}!**`)
                 .setDescription(`Здесь Вы можете получить поддержку от разработчика, задать свой вопрос, получить актуальную информацию по боту или просто приятно провести своё время! Ознакомьтесь с гидом ниже, чтобы не нарушать правила поведения в текстовых и голосовых каналах!`)
                 .addField(`\u200B`, `\u200B`)
                 .addField(`Основной канал:`, `__<#740560918613721103>__`, true)
@@ -87,8 +99,7 @@ bot.on("message", async message => {
                 .addField(`**Приглашение:**`, `__discord.gg/AbFExRq__`, true)
                 .addField(`**Сайт:**`, `__https://jeelanga.net/__`, true)
                 .addField(`**Разработчик:**`, `__<@!432085389948485633>__`, true)
-                .addField(`\u200B`, `\u200B`)
-                .setImage(`https://media.discordapp.net/attachments/634048309568208907/742374141075652738/kiss_136.gif`);
+                .setFooter(`Сообщение приветствия`, guild.iconURL({size: 4096, dynamic: true}));
 
             return bot.channels.cache.get(process.env.WelcomeChannel).send(WelcomeEmbed);
         }
@@ -96,11 +107,11 @@ bot.on("message", async message => {
         if(message.content.toLowerCase() === "set-ticket"){
             message.delete();
 
-            const TicketEmbed = new Discord.MessageEmbed()
+            const TicketEmbed = new MessageEmbed()
                 .setColor(process.env.EmbedGreen)
                 .setTitle(`**Создание тикета для связи с поддержкой**`)
                 .setDescription(`Нажав на реакцию - Вы создатите приватный канал, в котором сможете получить помощь от разработчика!`)
-                .setFooter(`Jeelanga поддержка`, group.iconURL({size: 4096, dynamic: true}));
+                .setFooter(`Jeelanga поддержка`, guild.iconURL({size: 4096, dynamic: true}));
 
             return bot.channels.cache.get(process.env.SupportChannel).send(TicketEmbed)
                 .then(msg => {
@@ -111,8 +122,9 @@ bot.on("message", async message => {
         if(message.content.toLowerCase() === "set-rules"){
             message.delete();
 
-            const RulesEmbed = new Discord.MessageEmbed()
+            const RulesEmbed = new MessageEmbed()
                 .setColor(process.env.EmbedGreen)
+                .setThumbnail(guild.iconURL({size: 4096, dynamic: true}))
                 .setTitle(`**Правила Discord-сервера**`)
                 .setDescription(`
                 **1:** Запрещено оскорблять участников Discord-сервера в любом виде.\n
@@ -126,24 +138,24 @@ bot.on("message", async message => {
                 \n
                 **НЕ ПРИКРЫВАЙТЕСЬ ИЗЪЯНАМИ ПРАВИЛ!**
                 `)
-                .setImage(`https://media.discordapp.net/attachments/634048309568208907/742493229844529244/TatteredSecondaryBlackbear-size_restricted.gif`);
+                .setFooter(``, guild.iconURL({size: 4096, dynamic: true}));
 
             return bot.channels.cache.get(process.env.RulesChannel).send(RulesEmbed);
         }
     }
 
-    if(message.author.id == process.env.OwnerID) return;
+    if(message.author.id == process.env.OwnerID || message.member.roles.cache.has(process.env.PremiumRole)) return;
     if(Words.wh_word.some(word => messages.includes(word))) return;
     if(Words.bad_word.some(word => messages.includes(word))){
         message.delete();
 
-        const BadWord = new Discord.MessageEmbed()
+        const BadWord = new MessageEmbed()
             .setColor(process.env.EmbedRed)
             .setTimestamp()
             .setThumbnail(message.author.displayAvatarURL({size: 4096, dynamic: true}))
             .setTitle(`${bot.user.username} » Запрещённое слово!`)
             .setDescription(`${message.author}, использовать такие слова запрещено!`)
-            .setFooter(`${bot.user.username}`, group.iconURL({size: 4096, dynamic: true}));
+            .setFooter(bot.user.username, guild.iconURL({size: 4096, dynamic: true}));
 
         return message.channel.send(BadWord).then(msg => msg.delete({timeout: 60000}));
     }
@@ -177,7 +189,7 @@ bot.on("raw", async event => {
                                 await channel.lockPermissions();
                                 await channel.updateOverwrite(user.id, {"VIEW_CHANNEL": true})
                             }).then(() => {
-                                const DeleteTicket = new Discord.MessageEmbed()
+                                const DeleteTicket = new MessageEmbed()
                                     .setColor(process.env.EmbedRed)
                                     .setTimestamp()
                                     .setThumbnail(user.displayAvatarURL({size: 4096, dynamic: true}))    
@@ -200,7 +212,7 @@ bot.on("raw", async event => {
                                 await channel.lockPermissions();
                                 await channel.updateOverwrite(user.id, {"VIEW_CHANNEL": true})
                             }).then(() => {
-                                const DeleteTicket = new Discord.MessageEmbed()
+                                const DeleteTicket = new MessageEmbed()
                                     .setColor(process.env.EmbedRed)
                                     .setTimestamp()
                                     .setThumbnail(user.displayAvatarURL({size: 4096, dynamic: true}))    
@@ -222,7 +234,7 @@ bot.on("raw", async event => {
     }
 
     if(event.t === "MESSAGE_REACTION_ADD"){
-        if(channel.name === `${member.roles.cache.has(process.env.PremiumRole) ? "premium-" : "ticket-"}${user.id}`){
+        if(channel.name === `${a ? "premium-" : "ticket-"}${user.id}`){
             if(data.emoji.name === process.env.DeleteReaction){
                 channel.delete();
             }
